requirements.txt
streamlit==1.35.0
transformers==4.38.2
torch==2.2.1
pandas==2.2.2
openai-whisper==20231117
langdetect==1.0.9
numpy==1.26.4
filelock==3.13.1